Added condition flag clear methods.
Added a method that clears every condition flag of every object in token
manager back to 0 at the end of the current run. However, this means
that frames are now "stateless". between execution loops. If they will
remain this way or if every SET needs an accompanying IF NOT ... SET[0]
is still being decided.